requirements.txt
fastapi
uvicorn
python-multipart
cloudinary
python-dotenv
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
torch
git+https://github.com/openai/CLIP.git
pillow
requests
openai
rembg
onnxruntime
qdrant-client
modal
google-genai